Some progress on a more advanced Embed

+++ classes/setup.js
@@ -14,12 +14,9 @@ class Setup {
         Setup._instance = this;
     }
     
-    async quickSetup (guild, member) {
+    async quickSetup (guild, member, deleteData) {
         // Delete all reactables (emoji, channels, roles) before starting
-        // TO-DO: IMPORTANT!!
-        // This is fine for testing but don't actually do this later.
-        // Deleting the best-of channel is gonna cause so many people to lose it
-        await this.startSetupFromScratch(guild);
+        await this.startSetupFromScratch(guild, deleteData);
 
         // Create #best-of channel
         const bestOf = await this.createBestOfChannel(guild);
@@ -63,7 +60,7 @@ class Setup {
 		});
     }
 
-    async startSetupFromScratch(guild) {
+    async startSetupFromScratch(guild, deleteData) {
         // Delete pre-existing guild data
         await guildSchema.deleteMany({ guildId: guild.id });
 
@@ -71,28 +68,36 @@ class Setup {
         await reactableSchema.find({ guildId: guild.id}).then(async (reactables) => {
             // Delete all channels associated with reactables
             for (const reactable of reactables) {
-                if (reactable.sendsToChannel) {
-                    const channel = guild.channels.cache.find(channel => channel.id === reactable.sendsToChannel);
-                    if (channel) {
-                        await channel.delete();
-                    }
-                }
 
-                for (const roleId of reactable.lockedBehindRoles) {
-                    const role = guild.roles.cache.find(role => role.id === roleId);
-                    if (role) {
-                        await role.delete();
-                    }
-                }
+                // TO-DO: We should check if Reto created these
+                // channels / emoji / roles before deleting them
+                // (fetch audit logs?)
 
-                for (const emojiId of reactable.emojiIds) {
-                    const emoji = guild.emojis.cache.find(emoji => emoji.id === emojiId);
-                    if (emoji) {
-                        try {
-                            await emoji.delete();
-                        } catch (error) {
-                            // This sometimes errors out "Unknown Emoji". No idea why that is, but it shouldn't be a problem.
-                            console.log(error);
+                if (deleteData) {
+                    if (reactable.sendsToChannel) {
+                        const channel = guild.channels.cache.find(channel => channel.id === reactable.sendsToChannel);
+                        if (channel) {
+                            await channel.delete();
+                        }
+                    }
+    
+                    for (const roleId of reactable.lockedBehindRoles) {
+                        const role = guild.roles.cache.find(role => role.id === roleId);
+                        if (role) {
+                            // TO-DO: This doesn't appear to work.
+                            await role.delete();
+                        }
+                    }
+    
+                    for (const emojiId of reactable.emojiIds) {
+                        const emoji = guild.emojis.cache.find(emoji => emoji.id === emojiId);
+                        if (emoji) {
+                            try {
+                                await emoji.delete();
+                            } catch (error) {
+                                // This sometimes errors out "Unknown Emoji". No idea why that is, but it shouldn't be a problem.
+                                console.log(error);
+                            }
                         }
                     }
                 }
